move sensor process to its own package
