+++ viam-cartographer_internal_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"testing"
-	"time"
 
 	"github.com/golang/geo/r3"
 	"github.com/pkg/errors"
@@ -17,7 +16,6 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/structpb"
 
-	cartofacade "github.com/viamrobotics/viam-cartographer/cartofacade"
 	inject "github.com/viamrobotics/viam-cartographer/internal/inject"
 )
 
@@ -398,77 +396,4 @@ func TestGetInternalStateEndpoint(t *testing.T) {
 			test.That(t, callback, test.ShouldBeNil)
 		})
 	})
-}
-
-func TestAddSensorReadingOffline(t *testing.T) {
-	svc := &cartographerService{Named: resource.NewName(slam.API, "test").AsNamed()}
-	svc.dataRateMs = 200
-	cf := cartofacade.Mock{}
-	buf := bytes.NewBuffer([]byte("12345"))
-	cf.AddSensorReadingFunc = func(
-		ctx context.Context,
-		timeout time.Duration,
-		sensorName string,
-		currentReading []byte,
-		readingTimestamp time.Time,
-	) error {
-		return nil
-	}
-	svc.cartofacade = &cf
-
-	// When addSensorReading returns successfully, no infinite loop
-	addSensorReadingOffline(context.Background(), svc, time.Now(), buf, 10*time.Second)
-
-	cf.AddSensorReadingFunc = func(
-		ctx context.Context,
-		timeout time.Duration,
-		sensorName string,
-		currentReading []byte,
-		readingTimestamp time.Time,
-	) error {
-		return errors.New("cant acquire lock")
-	}
-	svc.cartofacade = &cf
-
-	// When addSensorReading is erroring and the context is cancelled, no infinite loop
-	cancelCtx, cancelFunc := context.WithCancel(context.Background())
-	cancelFunc()
-	addSensorReadingOffline(cancelCtx, svc, time.Now(), buf, 10*time.Second)
-}
-
-func TestAddSensorReadingOnline(t *testing.T) {
-	svc := &cartographerService{Named: resource.NewName(slam.API, "test").AsNamed()}
-	svc.dataRateMs = 200
-	cf := cartofacade.Mock{}
-	buf := bytes.NewBuffer([]byte("12345"))
-	cf.AddSensorReadingFunc = func(
-		ctx context.Context,
-		timeout time.Duration,
-		sensorName string,
-		currentReading []byte,
-		readingTimestamp time.Time,
-	) error {
-		time.Sleep(1 * time.Second)
-		return nil
-	}
-	svc.cartofacade = &cf
-
-	// When addSensorReading blocks for longer than the data rate
-	timeToSleep := addSensorReadingOnline(context.Background(), svc, time.Now(), buf, 10*time.Second)
-	test.That(t, timeToSleep, test.ShouldEqual, 0)
-
-	cf.AddSensorReadingFunc = func(
-		ctx context.Context,
-		timeout time.Duration,
-		sensorName string,
-		currentReading []byte,
-		readingTimestamp time.Time,
-	) error {
-		return nil
-	}
-	svc.cartofacade = &cf
-
-	// When addSensorReading does not block for longer than the data rate
-	timeToSleep = addSensorReadingOnline(context.Background(), svc, time.Now(), buf, 10*time.Second)
-	test.That(t, timeToSleep, test.ShouldNotEqual, 0)
 }
